+++ pandas/tests/frame/methods/test_map.py
@@ -111,7 +111,8 @@ def test_map_na_ignore(float_frame):
     strlen_frame_na_ignore = float_frame_with_na.map(
         lambda x: len(str(x)), na_action="ignore"
     )
-    strlen_frame_with_na = strlen_frame.copy()
+    # Set float64 type to avoid upcast when setting NA below
+    strlen_frame_with_na = strlen_frame.copy().astype("float64")
     strlen_frame_with_na[mask] = pd.NA
     tm.assert_frame_equal(strlen_frame_na_ignore, strlen_frame_with_na)
